+++ src/Entity/WorkflowPlace.php
@@ -18,7 +18,7 @@ class WorkflowPlace
     #[ORM\Id]
     #[ORM\GeneratedValue]
     #[ORM\Column(type: Types::INTEGER)]
-    private ?int $id = null;
+    protected ?int $id = null;
 
     #[ORM\ManyToOne(targetEntity: WorkflowDefinition::class, inversedBy: 'places')]
     #[ORM\JoinColumn(nullable: false, onDelete: 'CASCADE')]